content: add colophon page at
This change involved making some changes to the build system to enable
injecting Git information into the build.

In local contexts, Hugo will use it's `GitInfo`, but that isn't
available during the nix build where `.git` is missing, so we replace
those sections with the current revision of the flake.

Also simplify the versioning to use short hashes.